+++ changelog.md
@@ -31,6 +31,7 @@
 
 ### Misc
 
+* Improved - Use fake player to place blocks.
 * Improved - The screen pf Flexible Marker now shows relative position.
 * Changed - Flexible Marker can now select up to 256 blocks.
 * Use conventional tags in recipe.