+++ changelog.md
@@ -1,3 +1,9 @@
+## 1.20
+
+### Misc
+
+* Update for 1.20.1
+
 ## 1.19
 
 ### Changes